Add VoltSigner to to Tools
I circulated this on April 30 on the dev mailing list. There was no reply, so I hope it's OK to move to the next step.

We’ve got a new tool we’d like to share with the Cordova community. It facilitates the creation of signing certificates for Android and iOS.

https://voltsigner.com

- It’s free
- No login is required
- We’re in the process of making of making the source public on GitHub
- Runs completely in the browser
- Generates Android and iOS .p12 files on a Windows or MacOS computers.
- Certificates are made 100% locally - passwords and files are not shared to the cloud.